// src/app/models/customer.ts
export class Customer {
  CustomerId: number;
  FirstName: string;
  LastName: string;
  Role: number;
  FullName: string;
  AddressOne: string;
  AddressTwo: string;
  Town: string;
  Country: string;
  RegionId: number;
  Region: string;
  ZipCode: string;
  Mobile: number;
  CreditCard: number;
  Email: string;
  Password: string;
  name: string;
  customer_id: number;
  partner_name: any;
  PartnerId: any;
}